+++ compiler-cli/src/docs.rs
@@ -4,6 +4,7 @@ use camino::{Utf8Path, Utf8PathBuf};
 
 use crate::{cli, hex::ApiKeyCommand, http::HttpClient};
 use gleam_core::{
+    analyse::TargetSupport,
     build::{Codegen, Mode, Options, Package},
     config::{DocsPage, PackageConfig},
     error::Error,
@@ -71,6 +72,7 @@ pub fn build(options: BuildOptions) -> Result<()> {
             target: None,
             codegen: Codegen::All,
             warnings_as_errors: false,
+            root_target_support: TargetSupport::Enforced,
         },
         crate::build::download_dependencies()?,
     )?;
@@ -156,6 +158,7 @@ impl PublishCommand {
 
         let mut built = crate::build::main(
             Options {
+                root_target_support: TargetSupport::Enforced,
                 warnings_as_errors: false,
                 codegen: Codegen::All,
                 mode: Mode::Prod,